[refactor] 프로젝트 구조 변경 - src 에 컴포넌트 및 page 관련 코드 이동 - next js 에서 public 은 root 에 있어야 해서 root 에 유지
