--- client/__tests__/index.test.tsx
@@ -1,14 +1,0 @@
-import { render, screen } from '@testing-library/react';
-import Home from '@/pages/index';
-
-describe('Home', () => {
-  it('renders a heading', () => {
-    render(<Home />);
-
-    // const heading = screen.getByRole('heading', {
-    //   name: /welcome to next\.js!/i,
-    // });
-
-    // expect(heading).toBeInTheDocument();
-  });
-});
