--- client/pages/index.tsx
@@ -1,5 +1,0 @@
-import Home from './Home';
-
-export default function HomeIndex() {
-  return <Home />;
-}
