+++ client/src/components/Header/Header.test.tsx
@@ -1,10 +1,10 @@
 import { render } from '@testing-library/react';
 
-import HomePage from './';
+import Header from '.';
 
-describe('HomePage', () => {
-  it('render header', () => {
-    const { container } = render(<HomePage />);
+describe('Header', () => {
+  it('render', () => {
+    const { container } = render(<Header />);
 
     expect(container.innerHTML).toContain('<img alt="logo"');
     expect(container).toHaveTextContent('블로그');
