+++ client/src/components/Posts/styles.ts
