+++ client/src/pages/_app.tsx
@@ -3,7 +3,7 @@ import type { AppProps } from 'next/app';
 
 import { Global } from '@emotion/react';
 
-import global from '../styles/global';
+import global from '@/styles/global';
 
 export default function MyApp({ Component, pageProps }: AppProps) {
   return (
